wake-lock/wakelock-supported-by-feature-policy.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('wake-lock', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise wake-lock.');
</script>
